+++ lib/services/local/presence_test.go
@@ -42,159 +42,6 @@ import (
 	"github.com/gravitational/teleport/lib/services/suite"
 )
 
-func TestRemoteClusterCRUD(t *testing.T) {
-	t.Parallel()
-	ctx := context.Background()
-
-	bk, err := lite.New(ctx, backend.Params{"path": t.TempDir()})
-	require.NoError(t, err)
-	t.Cleanup(func() { require.NoError(t, bk.Close()) })
-
-	presenceBackend := NewPresenceService(bk)
-	clock := clockwork.NewFakeClockAt(time.Now())
-
-	originalLabels := map[string]string{
-		"a": "b",
-		"c": "d",
-	}
-
-	rc, err := types.NewRemoteCluster("foo")
-	require.NoError(t, err)
-	rc.SetConnectionStatus(teleport.RemoteClusterStatusOffline)
-	rc.SetLastHeartbeat(clock.Now())
-	rc.SetMetadata(types.Metadata{
-		Name:   "foo",
-		Labels: originalLabels,
-	})
-
-	src, err := types.NewRemoteCluster("bar")
-	require.NoError(t, err)
-	src.SetConnectionStatus(teleport.RemoteClusterStatusOnline)
-	src.SetLastHeartbeat(clock.Now().Add(-time.Hour))
-
-	// create remote clusters
-	err = presenceBackend.CreateRemoteCluster(rc)
-	require.NoError(t, err)
-	err = presenceBackend.CreateRemoteCluster(src)
-	require.NoError(t, err)
-
-	// get remote cluster make sure it's correct
-	gotRC, err := presenceBackend.GetRemoteCluster("foo")
-	require.NoError(t, err)
-	require.Equal(t, "foo", gotRC.GetName())
-	require.Equal(t, teleport.RemoteClusterStatusOffline, gotRC.GetConnectionStatus())
-	require.Equal(t, clock.Now().Nanosecond(), gotRC.GetLastHeartbeat().Nanosecond())
-	require.Equal(t, originalLabels, gotRC.GetMetadata().Labels)
-
-	updatedLabels := map[string]string{
-		"e": "f",
-		"g": "h",
-	}
-
-	// update remote clusters
-	rc.SetConnectionStatus(teleport.RemoteClusterStatusOnline)
-	rc.SetLastHeartbeat(clock.Now().Add(time.Hour))
-	rc.SetMetadata(types.Metadata{
-		Name:   "foo",
-		Labels: updatedLabels,
-	})
-	err = presenceBackend.UpdateRemoteCluster(ctx, rc)
-	require.NoError(t, err)
-
-	src.SetConnectionStatus(teleport.RemoteClusterStatusOffline)
-	src.SetLastHeartbeat(clock.Now())
-	err = presenceBackend.UpdateRemoteCluster(ctx, src)
-	require.NoError(t, err)
-
-	// get remote cluster make sure it's correct
-	gotRC, err = presenceBackend.GetRemoteCluster("foo")
-	require.NoError(t, err)
-	require.Equal(t, "foo", gotRC.GetName())
-	require.Equal(t, teleport.RemoteClusterStatusOnline, gotRC.GetConnectionStatus())
-	require.Equal(t, clock.Now().Add(time.Hour).Nanosecond(), gotRC.GetLastHeartbeat().Nanosecond())
-	require.Equal(t, updatedLabels, gotRC.GetMetadata().Labels)
-
-	gotRC, err = presenceBackend.GetRemoteCluster("bar")
-	require.NoError(t, err)
-	require.Equal(t, "bar", gotRC.GetName())
-	require.Equal(t, teleport.RemoteClusterStatusOffline, gotRC.GetConnectionStatus())
-	require.Equal(t, clock.Now().Nanosecond(), gotRC.GetLastHeartbeat().Nanosecond())
-
-	// get all clusters
-	allRC, err := presenceBackend.GetRemoteClusters()
-	require.NoError(t, err)
-	require.Len(t, allRC, 2)
-
-	// delete cluster
-	err = presenceBackend.DeleteRemoteCluster(ctx, "foo")
-	require.NoError(t, err)
-
-	// make sure it's really gone
-	err = presenceBackend.DeleteRemoteCluster(ctx, "foo")
-	require.Error(t, err)
-	require.ErrorIs(t, err, trace.NotFound("key /remoteClusters/foo is not found"))
-}
-
-func TestTrustedClusterCRUD(t *testing.T) {
-	t.Parallel()
-	ctx := context.Background()
-
-	bk, err := lite.New(ctx, backend.Params{"path": t.TempDir()})
-	require.NoError(t, err)
-	t.Cleanup(func() { require.NoError(t, bk.Close()) })
-
-	presenceBackend := NewPresenceService(bk)
-
-	tc, err := types.NewTrustedCluster("foo", types.TrustedClusterSpecV2{
-		Enabled:              true,
-		Roles:                []string{"bar", "baz"},
-		Token:                "qux",
-		ProxyAddress:         "quux",
-		ReverseTunnelAddress: "quuz",
-	})
-	require.NoError(t, err)
-
-	// we just insert this one for get all
-	stc, err := types.NewTrustedCluster("bar", types.TrustedClusterSpecV2{
-		Enabled:              false,
-		Roles:                []string{"baz", "aux"},
-		Token:                "quux",
-		ProxyAddress:         "quuz",
-		ReverseTunnelAddress: "corge",
-	})
-	require.NoError(t, err)
-
-	// create trusted clusters
-	_, err = presenceBackend.UpsertTrustedCluster(ctx, tc)
-	require.NoError(t, err)
-	_, err = presenceBackend.UpsertTrustedCluster(ctx, stc)
-	require.NoError(t, err)
-
-	// get trusted cluster make sure it's correct
-	gotTC, err := presenceBackend.GetTrustedCluster(ctx, "foo")
-	require.NoError(t, err)
-	require.Equal(t, "foo", gotTC.GetName())
-	require.True(t, gotTC.GetEnabled())
-	require.EqualValues(t, []string{"bar", "baz"}, gotTC.GetRoles())
-	require.Equal(t, "qux", gotTC.GetToken())
-	require.Equal(t, "quux", gotTC.GetProxyAddress())
-	require.Equal(t, "quuz", gotTC.GetReverseTunnelAddress())
-
-	// get all clusters
-	allTC, err := presenceBackend.GetTrustedClusters(ctx)
-	require.NoError(t, err)
-	require.Len(t, allTC, 2)
-
-	// delete cluster
-	err = presenceBackend.DeleteTrustedCluster(ctx, "foo")
-	require.NoError(t, err)
-
-	// make sure it's really gone
-	_, err = presenceBackend.GetTrustedCluster(ctx, "foo")
-	require.Error(t, err)
-	require.ErrorIs(t, err, trace.NotFound("key /trustedclusters/foo is not found"))
-}
-
 // TestApplicationServersCRUD verifies backend operations on app servers.
 func TestApplicationServersCRUD(t *testing.T) {
 	t.Parallel()
